refactor(keymap): Remove third left thumbkey
As I plan to use 34 keys keyboard I decided to break my dependency to the third
left thumbkey from the beginning. Previously it was marked as `&trans` but I
think it is much better for it to be marked as `&none` as there is no button
that I assign into it.