refactor(keymap): Flip `SQT` for `SEMI`
I find that in Colemak-DH actually puts `SEMI` in that position. Plus with my
emacs, vim and coding in general `SEMI` is much more logical choice. Although
for writing `SQT` made much more sense.
I could try make it as a `tap-dance` as the nature of usage of `SEMI` and `SQT`
are usually being tapped. maybe I'd try it in the future.